homework week2
