+++ src/components/Greeting.ts
@@ -1,0 +1,7 @@
+import React from 'react';
+
+export default function Greeting({ name }: {
+  name: string;
+}) {
+  return React.createElement('P', null, 'Hello,', name, '!');
+}
