--- src/components/Greeting.tsx
@@ -1,12 +1,0 @@
-export default function Greeting({ name }: {
-  name: string;
-}) {
-  return (
-    <p>
-      Hello,
-      {' '}
-      {name}
-      !
-    </p>
-  );
-}
